BroadleafCommerce/QA#4078 Remove internal SessionFixationProtectionFilter
