+++ common/src/main/resources/config/bc/common.properties
@@ -201,3 +201,5 @@ allow.product.parent.category.sorting=true
 spring.main.allow-bean-definition-overriding=true
 
 spring.cache.jcache.config=jcache:merged-xml-resource
+
+unsecured.request=false
