+++ core/broadleaf-profile-web/src/main/java/org/broadleafcommerce/profile/web/site/security/LogUnsecuredSecureRequestFilter.java
@@ -1,0 +1,68 @@
+/*
+ * #%L
+ * BroadleafCommerce Profile Web
+ * %%
+ * Copyright (C) 2009 - 2016 Broadleaf Commerce
+ * %%
+ * Licensed under the Broadleaf Fair Use License Agreement, Version 1.0
+ * (the "Fair Use License" located  at http://license.broadleafcommerce.org/fair_use_license-1.0.txt)
+ * unless the restrictions on use therein are violated and require payment to Broadleaf in which case
+ * the Broadleaf End User License Agreement (EULA), Version 1.1
+ * (the "Commercial License" located at http://license.broadleafcommerce.org/commercial_license-1.1.txt)
+ * shall apply.
+ * 
+ * Alternatively, the Commercial License may be replaced with a mutually agreed upon license (the "Custom License")
+ * between you and Broadleaf Commerce. You may not use this file except in compliance with the applicable license.
+ * #L%
+ */
+package org.broadleafcommerce.profile.web.site.security;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.springframework.core.env.Environment;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.GenericFilterBean;
+
+import java.io.IOException;
+
+import javax.annotation.Resource;
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/**
+ * Filter used to protected against use unsecured http protocols.
+ *
+ */
+@Component("blLogUnsecuredSecureRequestFilter")
+public class LogUnsecuredSecureRequestFilter extends GenericFilterBean {
+
+    private static final Log LOG = LogFactory.getLog(LogUnsecuredSecureRequestFilter.class);
+
+
+    @Resource
+    protected Environment environment;
+
+
+    @Override
+    public void doFilter(ServletRequest sRequest, ServletResponse sResponse, FilterChain chain) throws IOException, ServletException {
+        HttpServletRequest request = (HttpServletRequest) sRequest;
+        HttpServletResponse response = (HttpServletResponse) sResponse;
+
+        if (SecurityContextHolder.getContext() == null) {
+            chain.doFilter(request, response);
+        }
+
+        if (!Boolean.parseBoolean(environment.getProperty("unsecured.request")) && !request.isSecure()){
+            LOG.info("Request is not secure for http. Use spring SessionFixationProtectionStrategy");
+        }
+
+        chain.doFilter(request, response);
+    }
+
+
+}
